+++ concordium-consensus/src/Concordium/Scheduler/Environment.hs
@@ -385,8 +385,8 @@ class
     runPLT ::
         (PVSupportsPLT (MPV m)) =>
         Token.TokenIndex ->
-        (forall m1. (Monad m1, PLTKernelPrivilegedUpdate m1, PLTKernelFail e m1, PLTAccount m1 ~ AccountIndex) => m1 a) ->
-        m (Either e a)
+        (forall m1. (Monad m1, PLTKernelPrivilegedUpdate m1, PLTKernelFail e m1, PLTAccount m1 ~ (AccountIndex, AccountAddress)) => m1 a) ->
+        m (Either e a, [Event])
 
     -- | Create a new protocol-layer token with the given 'PLTConfiguration'.
     --