+++ concordium-consensus/tests/globalstate/GlobalStateTests/AccountsMigrationP6ToP7.hs
@@ -41,13 +41,11 @@ import Concordium.GlobalState.CooldownQueue
 import Concordium.GlobalState.DummyData
 import Concordium.GlobalState.Persistent.Account
 import qualified Concordium.GlobalState.Persistent.Account.MigrationState as MigrationState
-import Concordium.GlobalState.Persistent.Account.ProtocolLevelTokens
 import Concordium.GlobalState.Persistent.Accounts
 import Concordium.GlobalState.Persistent.Bakers
 import Concordium.GlobalState.Persistent.BlobStore
 import Concordium.GlobalState.Persistent.BlockState
 import qualified Concordium.GlobalState.Persistent.BlockState.Modules as M
-import Concordium.GlobalState.Persistent.BlockState.ProtocolLevelTokens
 import Concordium.GlobalState.Persistent.Cooldown
 import qualified Concordium.GlobalState.Persistent.Trie as Trie
 import Concordium.ID.Types
@@ -88,31 +86,11 @@ dummyAccountEncryptedAmount =
         { _selfAmount = encryptAmountZeroRandomness dummyCryptographicParameters 10
         }
 
--- | A dummy-in memory token account state table.
-dummyTokenAccountStateTable :: (Hashed' TokenStateTableHash InMemoryTokenStateTable)
-dummyTokenAccountStateTable =
-    makeHashed $
-        InMemoryTokenStateTable
-            { inMemoryTokenStateTable =
-                Map.fromList
-                    [   ( TokenIndex 1,
-                          TokenAccountState
-                            { tasModuleState =
-                                Map.fromList
-                                    [ ("state_key1", "state_value1"),
-                                      ("state_key2", "state_value2")
-                                    ],
-                              tasBalance = TokenRawAmount 100
-                            }
-                        )
-                    ]
-            }
-
 -- | Create a test account with the given persisting data and stake.
 --  The balance of the account is set to 1 billion CCD (10^15 uCCD).
 testAccount ::
     forall av.
-    (IsAccountVersion av) =>
+    (IsAccountVersion av, SupportsPLT av ~ 'False) =>
     PersistingAccountData ->
     AccountStake av ->
     Transient.Account av
@@ -125,10 +103,7 @@ testAccount persisting stake =
           _accountReleaseSchedule = Transient.emptyAccountReleaseSchedule,
           _accountStaking = stake,
           _accountStakeCooldown = Transient.emptyCooldownQueue (accountVersion @av),
-          _accountTokenStateTable =
-            conditionally
-                (sSupportsPLT (accountVersion @av))
-                dummyTokenAccountStateTable
+          _accountTokenStateTable = CFalse
         }
 
 -- | Initial stake for a test account, set to 500 million CCD plus @2^accountIndex@ uCCD.