+++ concordium-consensus/tests/scheduler/SchedulerTests/Helpers.hs
@@ -789,6 +789,18 @@ assertFailureWithReason expectedReason result =
         [] -> assertFailure "No transaction failed"
         other -> assertFailure $ "Multiple transactions failed: " ++ show other
 
+-- | Assert the scheduler result has failed one chain update and check the reason.
+assertUpdateFailureWithReason :: Types.FailureKind -> SchedulerResult -> Assertion
+assertUpdateFailureWithReason expectedReason result =
+    case ftFailedUpdates $ srTransactions result of
+        [(_, reason)] ->
+            assertEqual
+                "The correct reason for failure is produced"
+                expectedReason
+                reason
+        [] -> assertFailure "No transaction failed"
+        other -> assertFailure $ "Multiple transactions failed: " ++ show other
+
 -- | Assert the scheduler have used energy the exact energy needed to deploy a provided V0 smart
 --  contract module. Assuming the transaction was signed with a single signature.
 --  The provided module should be a WASM module and without the smart contract version prefix.